We shouldn't be saving EDIT types for Shift Notifications, they're covered by the Shift Task Notifications
